Refactor HPA migration test and utility functions
- Moved the `toInt64` and `extractCPUAverageUtilization` functions from the HPA migration test file to the utils package for better reusability.
- Cleaned up the HPA migration test file by removing the now-externalized functions, improving readability and maintainability.
